pandas/pyproject.toml
[tool.poetry]
name = "duckdb-pd"
version = "0.1.0"
description = ""
authors = ["mehd-io <mehdi@mehd.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.1"
python-dotenv = "^1.0.0"
fsspec = "^2023.5.0"
s3fs = "^2023.5.0"
pyarrow = "^12.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
